Add support for the Simple Icons library
Motivation:

Lucide has decided not to cover brand icons. They recommend
using Simple Icons for this purpose. [1]

Many of the icons that we'd want to use on our Bento pages are brand
icons: GitHub, Reddit, Bluesky, etc.

This change starts work on supporting the Simple Icons library in Bento.

This will add a bit of complexity to the project (another integration to
maintain, more icons to choose from) but it satisfies this key need.

Details:

The change is backwards-compatible, as we assume Lucide Icons by
default and only require the new 'iconLibrary' config option when we
want to use that library.

I'm trying to do this in a way that it's easy to add support for even
more icon libraries. So, for anyone wanting to hack into this, they'll
be able to roll their own icons library too, with/out the existing ones.

The cost to performance is minimal, because we dynamically request the
icons that we need, instead of importing the whole library. I did that
in a previous version and it was 2.4 MB gzipped, 5.1 used! I believe
Lucide uses the same approach.

There's a bit of a FOUC while the image loads but I believe we can fix
this by displaying it as an <svg> tag (the Lucide way) instead of an
<img> one. That's my next step.

Also, these are colorful icons, but Simple Icons has a no-color CDN that
we should probably use. [2]

References:

[1] https://bsky.app/profile/joshwcomeau.com/post/3lbfgidke5c2s
[2] https://github.com/simple-icons/simple-icons?tab=readme-ov-file#cdn-usage